requirements.txt
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=0.24.2
joblib>=1.0.1
pyyaml>=5.4.1
dvc>=2.9.3
boto3>=1.18.0
jupyter>=1.0.0
pytest>=6.2.5
pytest-cov>=2.12.1
black>=21.7b0
flake8>=3.9.2
mypy>=0.910 